Refactor to use tryExecute
