+++ src/ClientDrawer.Bellissima/assets/src/datasources/clientdrawer.datasource.ts
@@ -1,6 +1,6 @@
 import { UmbControllerHost } from "@umbraco-cms/backoffice/controller-api";
 import { UmbDataSourceResponse } from "@umbraco-cms/backoffice/repository";
-import { tryExecuteAndNotify } from '@umbraco-cms/backoffice/resources';
+import { tryExecute } from '@umbraco-cms/backoffice/resources';
 import { V1Service } from '../api/sdk.gen';
 import { DataModel, HeaderActionModel } from "../api";
 
@@ -17,10 +17,10 @@ export class ClientDrawerDataSource implements IClientDrawerDataSource {
     }
 
     async getData(): Promise<UmbDataSourceResponse<DataModel>> {
-        return await tryExecuteAndNotify(this.#host, V1Service.getUmbracoClientdrawerApiV1Getdata())
+        return await tryExecute(this.#host, V1Service.getUmbracoClientdrawerApiV1Getdata())
     }
 
     async getHeaderActionData(): Promise<UmbDataSourceResponse<HeaderActionModel>> {
-        return await tryExecuteAndNotify(this.#host, V1Service.getUmbracoClientdrawerApiV1Getheaderactiondata())
+        return await tryExecute(this.#host, V1Service.getUmbracoClientdrawerApiV1Getheaderactiondata())
     }
 }
