+++ src/ClientDrawer.Bellissima/wwwroot/App_Plugins/ClientDrawer/assets.js
@@ -1,5 +1,5 @@
 import "@umbraco-cms/backoffice/auth";
-import { o as i } from "./index-B7egLMLd.js";
+import { o as i } from "./index-BuI3tihL.js";
 export {
   i as onInit
 };
